+++ .idea/deploymentTargetSelector.xml
@@ -4,7 +4,7 @@
     <selectionStates>
       <SelectionState runConfigName="app">
         <option name="selectionMode" value="DROPDOWN" />
-        <DropdownSelection timestamp="2024-12-27T23:20:44.573254Z">
+        <DropdownSelection timestamp="2025-03-27T17:59:09.335271Z">
           <Target type="DEFAULT_BOOT">
             <handle>
               <DeviceId pluginId="LocalEmulator" identifier="path=/Users/katielonsdale/.android/avd/Pixel_8_Pro_API_35.avd" />
