+++ app/src/main/java/com/katielonsdale/chatterbox/ui/components/PostCard.kt
@@ -30,12 +30,13 @@ import coil.request.ImageRequest
 import com.katielonsdale.chatterbox.R
 import com.katielonsdale.chatterbox.SampleData
 import com.katielonsdale.chatterbox.api.data.Post
-import com.katielonsdale.chatterbox.ui.home.CommentCard
+import com.katielonsdale.chatterbox.ui.components.CommentCard
 import sh.calvin.autolinktext.rememberAutoLinkText
 import java.time.ZonedDateTime
 import java.time.format.DateTimeFormatter
 
 
+// used for Newsfeed and Circle Newsfeed, only displays 2 comments
 @Composable
 fun PostCard(
     post: Post,
